fix: faq section cleaning
