+++ src/features/buyer/faq/components/Tabs.tsx
