+++ src/features/buyer/faq/components/TicketDetails.tsx
