+++ src/features/buyer/faq/components/quick-guide.tsx
